修复shader hash被fmodel删掉的问题

+++ Source/Ruri.FModelHook/Game/SBUE/ShaderDecompiler/Passes/Pass030_ScanMaterialPackages.cs
@@ -956,14 +956,14 @@ internal static class Pass030_ScanMaterialPackages
                 BaseIndex = parameter.BaseIndex,
                 Size = parameter.Size,
                 BufferIndex = parameter is FShaderResourceParameterInfo resource ? resource.BufferIndex : (byte)0,
-                Type = parameter is FShaderResourceParameterInfo typed ? typed.Type : (byte)0
+                Type = parameter is FShaderResourceParameterInfo typed ? (byte)typed.Type : (byte)0
             }).ToList() ?? new List<UnifiedShaderResourceParameterInfo>(),
             SRVs = parameterMapInfo.SRVs?.Select(parameter => new UnifiedShaderResourceParameterInfo
             {
                 BaseIndex = parameter.BaseIndex,
                 Size = parameter.Size,
                 BufferIndex = parameter is FShaderResourceParameterInfo resource ? resource.BufferIndex : (byte)0,
-                Type = parameter is FShaderResourceParameterInfo typed ? typed.Type : (byte)0
+                Type = parameter is FShaderResourceParameterInfo typed ? (byte)typed.Type : (byte)0
             }).ToList() ?? new List<UnifiedShaderResourceParameterInfo>(),
             LooseParameterBuffers = parameterMapInfo.LooseParameterBuffers?.Select(buffer => new UnifiedShaderLooseParameterBufferInfo
             {
